requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-binascii==2.0.4
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
bidict==0.23.1
binho-host-adapter==0.1.6
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
eventlet==0.38.2
Flask==3.1.0
Flask-SocketIO==5.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pyftdi==0.56.0
pyserial==3.5
python-engineio==4.11.1
python-socketio==5.12.0
pyusb==1.2.1
requests==2.32.3
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
simple-websocket==1.1.0
sysv-ipc==1.1.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wsproto==1.2.0
